+++ rento-api/src/main/java/com/kbe5/api/domain/company/controller/CompanyController.java
@@ -1,12 +1,12 @@
 package com.kbe5.api.domain.company.controller;
 
-import com.kbe5.api.domain.company.dto.request.CompanyBiznumberRequest;
-import com.kbe5.api.domain.company.dto.request.CompanyRegisterRequest;
-import com.kbe5.api.domain.company.dto.request.CompanyUpdateRequest;
-import com.kbe5.api.domain.company.dto.response.CompanyDeleteResponse;
-import com.kbe5.api.domain.company.dto.response.CompanyRegisterResponse;
-import com.kbe5.api.domain.company.dto.response.CompanyResponse;
-import com.kbe5.api.domain.company.dto.response.CompanyUpdateResponse;
+import com.kbe5.api.domain.company.service.dto.request.CompanyBiznumberRequest;
+import com.kbe5.api.domain.company.service.dto.request.CompanyRegisterRequest;
+import com.kbe5.api.domain.company.service.dto.request.CompanyUpdateRequest;
+import com.kbe5.api.domain.company.service.dto.response.CompanyDeleteResponse;
+import com.kbe5.api.domain.company.service.dto.response.CompanyRegisterResponse;
+import com.kbe5.api.domain.company.service.dto.response.CompanyResponse;
+import com.kbe5.api.domain.company.service.dto.response.CompanyUpdateResponse;
 import com.kbe5.common.response.api.ApiResponse;
 import org.springframework.http.ResponseEntity;
