+++ rento-api/src/main/java/com/kbe5/api/domain/company/service/dto/request/CompanyRegisterRequest.java
@@ -1,0 +1,13 @@
+package com.kbe5.api.domain.company.service.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record CompanyRegisterRequest(
+
+        @NotNull
+        int bizNumber,
+
+        @NotBlank
+        String name
+) {}
